{cmd, auth, utils}: Refactor & modularize parts of the code in preparation for the e2e tests (#154)
### What

Refactor and modularize export parts of the code in preparation for the e2e tests.

What was exported into reusable functions:
- http signature creator
- conversion helpers to make the following path easier: `txnbuild.Operation <> XDR <> base64 <> XDR <> txnbuild.Operation`

Also:
- Fixed the 500 crash that would prevent the feeBumpTx endpoint from working due to a missing dependency.
- Renamed the variable `wallet-signing-key` to `client-auth-public-keys`, which now is more descriptive and allows multiple signers to be set, making key rotation easier.

### Why

While writing the e2e tests, stumbled upon parts of the code that had only one way methods available. For instance, the code knows how to check a signature but doesn't know how to produce it. That led me to create such methods and export them in the `pkg` folder, so they can be used externally.