services: fix ingestor not catching-up (#173)
### What

Upgrade the ingestor triggers, so it can catch up to the latest ledger, if outdated.

### Why

The ingestor was depending oin the health check, that happens every ~5s, which is about the same period of the consensus. This means that if the service got downtime, it would never catch up to the latest ledger.