indexer: PushStateChanges will also push the operation and transaction (#247)
### What

When pushing the StateChange+Participant to the buffer, we will also push:
1. The Transaction+Participant
2. The Operation+Participant, if available

### Why

For contract transfers, not all participants are identified in the Operation parser (the destination of the transfer, for instance), so we gather it from the state changes associated with that operation.

This is consistent with what we do for the above tier pair: Transaction & Operations.

Relates to #218